Modificaciones recientes al proyecto en función de agregar reutilización basada en herencia.
