+++ FacturacionSoda/src/facturacionsoda/Jsonable.java
@@ -1,0 +1,14 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package facturacionsoda;
+
+/**
+ *
+ * @author Jheron Chacon
+ */
+public interface Jsonable{
+    public abstract void printToJson();
+}
