+++ FacturacionSoda/src/facturacionsoda/TipoIngrediente.java
@@ -1,24 +1,89 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package facturacionsoda;
 
-/**
- *
- * @author Jheron Chacon
- */
-public class TipoIngrediente {
-    private final static String VEGETAL = "Vegetal";
-    private final static String CARNE = "Carnes";
+import Mensajeria.ClasificacionTipoDeIngrediente;
+
+public class TipoIngrediente 
+{
+    private String nombre;
+    private String mensaje;
+    private boolean vegetal, carne, lacteo, licor, granos;
     
-    private boolean vegetal = false;
-    private boolean carne = false;
+    public TipoIngrediente()
+    {
+    }
     
-    public String getClasificacion(){
-        return this.vegetal ? TipoIngrediente.VEGETAL :
-                this.carne ? TipoIngrediente.CARNE :
+    public TipoIngrediente(String nombre)
+    {
+        this.nombre = nombre;
+    }
+    
+    public String getMensaje()
+    {
+        return this.mensaje;
+    }
+    
+    public String getNombre()
+    {
+        return this.nombre;
+    }
+    
+    public void setNombre(String nombre)
+    {
+        this.nombre = nombre;
+    }
+    
+    public void setCarne()
+    {
+        this.carne = true;
+        this.granos = false;
+        this.lacteo = false;
+        this.licor = false;
+        this.vegetal = false;
+    }
+    
+    public void setVegetal()
+    {
+        this.vegetal = true;
+        this.carne  = false;
+        this.granos = false;
+        this.lacteo = false;
+        this.licor = false;
+    }
+    
+    public void setLacteo()
+    {
+        this.lacteo = true;
+        this.carne = false;
+        this.granos = false;
+        this.licor = false;
+        this.vegetal = false;
+    }
+    
+    public void setLicor()
+    {
+        this.licor = true;
+        this.carne = false;
+        this.granos = false;
+        this.vegetal = false;
+        this.lacteo = false;
+    }
+    
+    public void setGranos()
+    {
+        this.granos = true;
+        this.carne = false;
+        this.lacteo = false;
+        this.licor = false;
+        this.vegetal = false;
+    }
+    
+    public String getClasificacion()
+    {
+        return this.vegetal ? ClasificacionTipoDeIngrediente.VEGETAL:
+                this.carne ? ClasificacionTipoDeIngrediente.CARNE:
+                this.granos ? ClasificacionTipoDeIngrediente.GRANOS:
+                this.lacteo ? ClasificacionTipoDeIngrediente.LACTEO:
+                this.licor ? ClasificacionTipoDeIngrediente.LICOR:
                 null;
     }
 }
